PixelPingu/Public/src/requirements.txt
flask
Werkzeug
torch
torchvision
Pillow
numpy
